+++ apps/demo/index.ts
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 import "dotenv/config";
 import {
   VercelAIAdapter,
@@ -8,6 +9,14 @@ import { openai } from "@ai-sdk/openai";
 import { generateText } from "ai";
 import PuzzletTypes, { Tools } from "./puzzlet1.types";
 import { FileLoader, createAgentMark } from "@agentmark/agentmark";
+=======
+import 'dotenv/config';
+import { createAgentMarkClient, VercelAIModelRegistry, VercelAIToolRegistry } from "@agentmark/vercel-ai-v4-adapter";
+import { openai } from '@ai-sdk/openai';
+import { generateObject } from 'ai';
+import { FileLoader } from '@agentmark/agentmark-core';
+import AgentmarkTypes, { Tools } from './agentmark.types';
+>>>>>>> main
 
 const modelRegistry = new VercelAIModelRegistry();
 const loader = new FileLoader("./fixtures");
@@ -20,30 +29,21 @@ const tools = new VercelAIToolRegistry<Tools>().register(
   ({ location }) => ({ tempC: 22 })
 );
 
-const adapter = new VercelAIAdapter<PuzzletTypes, typeof tools>(
-  modelRegistry,
-  tools
-);
-const agentMark = createAgentMark({
+const agentMark = createAgentMarkClient<AgentmarkTypes>({
   loader,
-  adapter,
+  modelRegistry,
+  toolRegistry: tools,
 });
 
-async function run() {
-  const prompt = await agentMark.loadTextPrompt("test/math2.prompt.mdx");
+async function run () {
+  const prompt = await agentMark.loadObjectPrompt('test/math2.prompt.mdx');
   const props = {
     userMessage: "Whats 2 + 3?",
   };
 
-  const vercelInput = await prompt.format({ props });
-  const result = await generateText(vercelInput);
-  result.toolResults?.forEach((toolResult) => {
-    console.log(toolResult);
-    if (toolResult.toolName === "weather") {
-      console.log(toolResult.result.tempC);
-    }
-  });
-  console.log(result.text);
+  const vercelInput = await prompt.format({props});
+  const result = await generateObject(vercelInput);
+  console.log(result.object.answer);
 }
 
 run();